IGNITE-21513 Fixed ActiveActorTest#testChangeLeaderForce is flaky
